Cleanup contract test Utils
